// src/app/models/user-profile-model.ts
export interface UserProfileDetails {
  userLevelId: number;
  userRoleId: number;
  regionId: number;
  countryId: number;
  companyId: number;
  partnerCompanyId: string;
  userName: string;
  firstName: string;
  middleName: string;
  lastName: string;
  sex: string;
  employeeNo: string;
  eMail: string;
  nationality: string;
  pEMail: string;
  ccMobile: string;
  acMobile: string;
  noMobile: string;
  ccPhone: string;
  acPhone: string;
  noPhone: string;
  ccDPhone: string;
  acDPhone: string;
  noDPhone: string;
  ccFax: string;
  acFax: string;
  noFax: string;
  status: string;
  userLevelName: string;
  userRoleName: string;
  regionName: string;
  countryName: string;
  state: string;
  countryISOCode: string;
  companyName: string;
  bussName: string;
  branchName: string;
  mainDepartmentName: string;
  sectionName: string;
  jobFunction: string;
  hoName: string;
  designationName: string;
  currencyCode: string;
  managerName: string;
}